requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-login==0.6.3
gunicorn==21.2.0
python-dotenv==1.0.1
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
# Database
SQLAlchemy==2.0.27
alembic==1.13.1
# LLM Integration
langchain==0.1.9
openai==1.12.0
requests>=2.31.0
# Content Management
markdown==3.5.2
python-frontmatter==1.1.0
# Security
PyJWT==2.8.0
bcrypt==4.1.2
# Testing
pytest==8.0.2
pytest-cov==4.1.0
# Caching
redis==5.0.1
flask-caching==2.1.0
# Task Queue
celery==5.3.6
# API Documentation
flasgger==0.9.7.1
# Image Processing
Pillow==10.2.0
# System Monitoring
psutil>=5.9.0 